requirements.txt
PyQt5>=5.15.0
Pillow>=10.0.0
mistralai>=0.0.12
keyboard>=0.13.5
python-dotenv>=1.0.0
markdown2>=2.4.0
